backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
pyjwt==2.8.0
python-multipart==0.0.6
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.7
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0